+++ Jolt/Math/Float4.h
@@ -5,7 +5,7 @@
 
 JPH_NAMESPACE_BEGIN
 
-/// Class that holds 4 float values. Convert to Vec4 to perform calculations.
+/// Class that holds 4 decimal values. Convert to Vec4 to perform calculations.
 class [[nodiscard]] Float4
 {
 public:
@@ -13,18 +13,18 @@ public:
 
 				Float4() = default; ///< Intentionally not initialized for performance reasons
 				Float4(const Float4 &inRHS) = default;
-				Float4(float inX, float inY, float inZ, float inW) : x(inX), y(inY), z(inZ), w(inW) { }
+				Float4(decimal inX, decimal inY, decimal inZ, decimal inW) : x(inX), y(inY), z(inZ), w(inW) { }
 
-	float		operator [] (int inCoordinate) const	
+	decimal		operator [] (int inCoordinate) const	
 	{ 
 		JPH_ASSERT(inCoordinate < 4); 
 		return *(&x + inCoordinate); 
 	}
 
-	float		x;
-	float		y;
-	float		z;
-	float		w;
+	decimal		x;
+	decimal		y;
+	decimal		z;
+	decimal		w;
 };
 
 static_assert(is_trivial<Float4>(), "Is supposed to be a trivial type!");
